接入sentinel

+++ settlement-service/src/main/java/com/dili/settlement/service/impl/SettleOrderServiceImpl.java
@@ -158,7 +158,7 @@ public class SettleOrderServiceImpl extends BaseServiceImpl<SettleOrder, Long> i
         output.setData(itemList);
         output.setPageNum(page.getPageNum());
         output.setPageSize(page.getPageSize());
-        output.setTotal(Long.valueOf(page.getTotal()).intValue());
+        output.setTotal(page.getTotal());
         output.setStartRow(page.getStartRow());
         output.setEndRow(page.getEndRow());
         output.setPages(page.getPages());
